Fix #83691: Viewport Tools Settings panel appears after reload
Per convention we never show the tool properties when the header is
invisible.
Opening a file where the header is hidden would show the tool properties
though.

So when changing the `RGN_TYPE_HEADER` visibility, everything is taken
care of for `RGN_TYPE_TOOL_HEADER`as well (its `RGN_FLAG_HIDDEN`is set,
`RGN_FLAG_HIDDEN_BY_USER` is left untouched though [this is only used
for hiding the tool properties "by hand", independent from the header it
self].

When saving/opening the file, we loose the state of `RGN_FLAG_HIDDEN` (a
part that admittedly I havent investigated deeply yet - it seems to
either get reset on load or not saved to file even?) and since
`RGN_FLAG_HIDDEN_BY_USER` is also not set, it just ends up being
visible.

This PR adds an additional check in some places that makes sure a
`RGN_TYPE_TOOL_HEADER` is not visible in case its corresponding
`RGN_TYPE_HEADER` is hidden.