Fix #129542: GPv3: Layers and Groups don't generate callback to message bus
This was a regression from 4.2.
The fix adds `WM_msg_publish_rna_prop` to places where the
layers/groups change or the active one changes.

Note: This does not cover every possible scenario and only fixes
the regressions from 4.2. The message bus is still very unreliable
when it comes to detecting changes of properties.