Fix #128712: Workbench: Crash dragging window to another screen
The crash was caused by the texture wrapper being out of date.
This happened after workbench finished rendering AntiAliasing and
going into blit-result-only mode. Doing so was causing a use after
free caught by ASAN.

This is a partial fix as workbench render is then becomes
uninitialized (black most likely) after dragging the window to the
second monitor.

Fixing this issue should be done in a separate commit as it is more
involved and too dangerous for 4.3.