Fix #125787: Relative texture paths do not work with brush assets
After the brush assets project was merged, image blocks that were
associated with new brush assets began to look in a different location
for the associated image file when using relative paths. With a main
blend file saved, this process would look at the essentials asset
library instead of the current blend file.

To fix this issue, this commit adds a number of overriden methods for
the IMAGE_OT_open operator to indicate how relative paths should be
processed when the associated data block is ultimately moved to a
separate .blend file.