Bug Reporting: Increased log details for Windows bug-reporting scripts
Previously, logs from Windows bug-reporting scripts would include a
minimal level of messages from Blender itself due to the usage of
the default log level.
With this change, the log level for these scripts is now set to the
maximum level, so all debug details from the Blender application
will be available in the generated log file, which users can
attach to bug reports.