Fix #130066: Crash after joining Grease Pencil objects.
The Join operator for Grease Pencil was making a shallow copy of the drawings
array, with the assumption that the source object would no longer be needed.
This is not the case, the source object is still in the blend data and still shares the same drawings pointer as the target object. This is invalid, the drawings should owned by one object.

To fix this the patch makes a deep copy of the drawings so the source and target have their own sets of drawings. Data is still shared at the CustomData level, so these copies are not very expensive.