WIP: Vulkan: Remove empty attachments
Some Vulkan platforms don't support framebuffers with gaps between the
color attachments. Workbench framebuffers can create gaps.
(`in_front_fb`, `main_fb`, `gbuffer_in_front_fb` and `gbuffer_fb`).

This PR implements a detection mechanism to detect gabsb in the
framebuffer. This is only active in builds with assert turned on and on
when the Vulkan backend is enabled. It also changes the mentioned
framebuffers to pass this test.

Subpass transitions are not an issue as the attachments are moved from
output to input.