Fix #130268: Assertion triggered when animate property from Material panel
Fix an assertion that an embedded data-block has a zero 'real user'
count. Apparently it's possible for the shader node tree (embedded by
the material) to have a user count of 1.

Since that looks valid to me (only one user, namely the material itself)
I think it's fine to extend the assertion to that.

I did keep the assertion, to ensure that the embedded data-block is not
shared by multiple users. That shouldn't be possible, in any case.