Fix: GPv3: Edit Mode: Crash during extrude
The crash was caused by the selection attribute not being
initialized properly.

The fix uses `ed::curves::ensure_selection_attribute` to ensure
the attribute is created and on the right domain.

Extrude should only work for a point selection, so the poll
function of the operator is updated to reflect this.