Platforms: X11 Implementation of getWindowUnderCursor
Unlike Windows and MacOS, X11 does not have a ghost implementation of
getWindowUnderCursor. This means it gets the default behavior for "what
is the topmost window at the mouse cursor location", which is to check
all windows in reverse order and return the first that contains the
location. This works okay in the most likely scenario, where newer
windows are positioned above older ones, but does not work when they
are reversed. A better result is needed for multi-window drag and drop
and for area maintenance routines like docking.  This PR includes a
version of getWindowUnderCursor specifically for X11.