Core: Fix Incorrect FP Precision With "None" Units
Float display when scene units are set to "None" is currently
incorrect. When set to other unit types we correctly show a maximum
number of digits of precision. But set to none we display with that
number of decimal places instead. For values above 100 or so that means
we display a value that is greater than the available precision, which
is confusing for users. This PR changes the formatting from decimals to
precision, and also increases the displayed precision to 6 places which
is our current maximum.