Core: Increase Float Precision to Seven Digits
This PR increases the maximum number of significant digits we display
for floating point values from 6 to 7. This is the amount of precision
with IEEE 754 single precision floats. 23 bits for mantissa plus some
"hidden" bit magic because of normalization and therefore 7.22 decimal
digits of precision. With this one extra digit users have their
accuracy increased 10X.