UI: Enable Node name header before all other buttons
The Node name (in header) is wider that the available space,
and when button/icon is displayed on right side of Node,
it is not working correctly.

For example for Viewer Geometry Node, the left side of the
button is not working correctly.

By changing the order of displaying/enabling, the buttons/icons
are fixes, and it is working correctly.

Fixes: #130244