Fix #130353: GPv3: Fill strokes disappear when using bucket fill
The issue was that when the fill strokes were merged into the drawing,
the `fill_opacity` attribute would be created on the drawing
(when it didn't exist before) and initialized to 0. This meant that
all previously drawn fill strokes would disappear.

The fix manaully initializes the `fill_opacity` attribute on the
target geometry before calling `join_geometries`.

Ideally this could be something the attribute system handles
or the `join_geometries` function takes care of by getting
some information about how to initialize attributes.