Fix #130368: GPv3: Canvas overlay incorrect offset
The issue was that the evaluated object doesn't have an active layer
This causes the canvas to be offset twice by the object position.

The reason was that the code scaled the whole transformation
matrix by two to match the code in 4.2, but then had to overwrite
the location of the transformation again to counteract the scaling.

The fix is to not counteract the scaling and just scale the 3x3 part
of the matrix instead. This way we can remove the part afterwards
that writes to the location of the transform.