Cleanup: Remove unused variables in sculpt_undo.cc
Previously, the undo type was stored at the node level, so undo pushes
inside sculpt mode that didn't need nodes still appended this data so
that they could be appropriately filtered out. This commit removes
instances where this was the case and clarifies other checks that used
the vector of undo nodes.