GPU: Move vertex format to enum class
The goal of this PR is to fuse the enum for
vertex format and texture data formats
(not texture format).

This allows to have more generic utility functions
inside the gpu backend since the attribute types
can be casted to `DataType`, deduplicating logic
in some place.

This PR only contains renaming.