WIP: Tools: Add cleaner of bpy install folder for wheel
The goal of this cleaner is to replace all libraries that are symbolic
link with a regular files, and remove all library files that are not
actually used.

This circumvents the situation when wheel packing converts symbolic
links (libraries with ABI in their names) to regular files due to the
limitation of the zip archives. Doing a manual cleanup pass allows to
run regression tests after the cleanup is done. It is implemented as
as standalone tool so that it can be invoked when it is really needed.

In practice this lowers Linux bpy wheel file size from 540 MB down to
around 362 MB.