Fix #130570: Grease Pencil use average pressure to fill stroke
When using a pressure sensitive tablet to draw strokes using Fill tool
(with `Shift` key held down so it generates fill), grease pencil used to
take the first ever pressure event to determine the opacity of the fill,
this is counter intuitive and hard to use. This patch changes the
behaviour to use average pressure along the stroke to determine fill
opacity, making it much more intuitive and easier to control.