UI: Keymap: Show alert when keymap operator not found
At the moment there is no way to tell if a keymap is being assigned to
a valid operator or not. This can be especially problematic when
migrating preferences and some operators no longer exist.

Alert the user when a keymap item is not available, by showing an icon
indicator, label, and marking the text input field red.

This is probably far from the most efficient way of doing it, if you
have ideas on how to improve it I'm open.

Hopefully we can make feedback part of a future re-design of the
entire keymap editor. For the time being, it's a small band-aid to
help users through the transition.