Anim: rename `.action_slots` to `.action_suitable_slots`
`AnimData`, NLA strips, and action constraints all have an
`action_slots` field in RNA. The purpose of this field is to list
the slots of the currently assigned action (if any) that are suitable
for the relevant ID. However, this is not clear from the naming.
In particular, given that there is another field `action_slot` that
represents the currently assigned slot, the name `action_slots` could be
easily misinterpreted as a way to assign multiple slots at once, which
is not possible.

To help clarify its actual purpose and meaning, this PR renames the
field to `action_suitable_slots`.

As a bonus, this also ends up decluttering the Python autocomplete when
looking for things related to `action_slot`.