UI: Show File Name on Recent Item Tooltip
When the recent files list is shown on the Splash screen, the
horizontal is constrained and so longer items will have their names
truncated. Unfortunately the popup tooltip shows the path but not
file name, so doesn't help in this situation. This PR adds the file
name as the first item in the tooltip. This was requested on RCS:
https://blender.community/c/rightclickselect/xVQE