Fix incorrect origin translation of View3D projection in ortho mode.
This was part of a fix for #130635, split out into a separate PR
because it affects a common editor function. The issue is that with a
combination of:
1. Orthographic view/camera mode
2. Camera "shift" (offset of the view from the central axis)

In this case the translation part of the perspective matrix (`persmat`/`persinv` in `RegionView3D`) contains
the center of the actual view rectangle, while the view matrix (`viewmat`/`viewinv`) contains the unshifted view direction. These vectors are the same when the view or camera has no shift. In the case of the Grease Pencil fill tool the shift is caused by defining an "image region".

The `ED_view3d_win_to_3d` function computes a ray-cast origin based on the perspective X/Y axes, but then uses the viewmat translation as the offset, ignoring any camera shift.