Fix #130707: Weight paint gradient tool doesnt fill behind start point
See blender/blender-assets!15 for the actual asset file change.

As part of the brush asset project, various weight paint brushes had
their falloff curve changed. This had the unintended effect of changing
the behavior of the gradient tool, as it shares the currently active
brush's settings.

This behavior is unintuitive and not well documented, as an interim fix,
the bundled assets were reverted.