Fix #131067: Stencil remains in view with tools that don't use it
This is more associated with brush assets. When non-brush tool is
selected, active brush in that context is not changed (see:
`toolsystem_ref_link`).
Stencil is drawn if brush has active texture on
it (`paint_draw_tex_overlay`), since the `Brush` is still referenced to
old value, stencil is still drawn. To fix this, set `pcontext->brush=nullptr`
if active tool is not brush.

Also note that, trim and filter tools are not brushes but
they intetionally draws brush cursor e140f263d0.